requirements.txt
# ConcursIA Backend - Dependências
# Framework web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Banco de dados
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
alembic==1.13.1

# Validação e configuração
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0
pydantic[email]

# Autenticação
PyJWT==2.11.0

# Utilitários
python-dotenv==1.0.0
python-multipart==0.0.6

# RAG (para integração com Chatbot)
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.7
langchain-core==0.2.38
langchain-huggingface==0.0.3
langgraph==0.2.16
chromadb==0.4.22
transformers>=4.39.0
sentence-transformers>=2.6.0

# Testes
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
